Fixes connection time out errors in Cloud Storage upload
Anecdotally I this is the main source of pipeline failures at the moment. When uploading large files, the GCS API breaks the file into 100 MB chunks, and opens a connection to complete each chunk. Connections last 1 minute, so you need to be able to sustain 100MB/minute upload for the duration of your upload, which I suspect Miranda is not always managing at the moment, especially with concurrent pipelines uploading data.

This PR sets the chunk size on uploaded blobs to 10MB. Testing locally on a throttled network connection, this fixes the problem for me. If the code looks good I'd like to test on a pipeline for a day or two and observe if the failure rate does indeed go down in practice before merging.