Removes uuid table migration support
Attempting to use this at the moment would always cause the assertion error "Not safe to use as not yet protected against readtimeouts". Now that all our active projects use the firestore uuid table (and have done for quite some time), we can remove this rather than trying to fix it.